updated select style in panel
