+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/utils.tsx
@@ -6,8 +6,6 @@ import {
   FormControlLabelProps,
   Radio,
   RadioProps,
-  Select,
-  SelectProps,
   useRadioGroup,
   withStyles,
 } from '@material-ui/core';
@@ -111,14 +109,6 @@ export const StyledCheckboxLabel = withStyles({
     />
   ),
 );
-
-export const StyledSelect = withStyles({
-  root: {
-    '&:focus': {
-      backgroundColor: 'transparent',
-    },
-  },
-})((props: SelectProps) => <Select {...props} />);
 
 const AACategoryPhaseMap: { [key: string]: any } = {
   na: {
